Create the Asteroid Object
This commit changes how the asteroid moves on the screen. Right now,
the asteroid method 'update' only needs to know if it is moving forward
and does not take into consideration whether the game model wants it to
rotate. This method helps the asteroid to move on the screen by updating
it's coordinates according to delta-time.